+++ destinations/faros-destination/test/converters/harness.test.ts
@@ -6,7 +6,7 @@ import os from 'os';
 import pino from 'pino';
 
 import {InvalidRecordStrategy} from '../../src';
-import {initMockttp, tempConfig} from '../temp';
+import {initMockttp, tempConfig} from '../testing-tools';
 import {CLI, read} from './../cli';
 import {harnessAllStreamsLog} from './data';
