+++ destinations/faros-destination/test/converters/phabricator.test.ts
@@ -4,7 +4,7 @@ import _ from 'lodash';
 import {getLocal} from 'mockttp';
 import pino from 'pino';
 
-import {initMockttp, tempConfig} from '../temp';
+import {initMockttp, tempConfig} from '../testing-tools';
 import {CLI, read} from './../cli';
 import {phabricatorAllStreamsLog} from './data';
